Port console decoder wrapper from C11 to C++23
Rename all .c → .cpp, fix void* implicit casts, replace memalign with
aligned_alloc, convert C11 _Atomic/stdatomic.h to std::atomic, wrap
Linux-only cpu_set_t/pthread_setaffinity_np in #ifdef __linux__, fix
AOM API call sites (aom_codec_av1_dx(), local iterator, correct
dec_cfg fields). Builds clean on macOS arm64 against libaom, 3/4
tests pass (job_queue has timer-granularity flakiness).

+++ aom-av1-prototype/verification_outputs/merge/test_job_queue.cpp
@@ -25,7 +25,7 @@ static void test_basic_capacity(void) {
     ASSERT(av1_frame_queue_init(&q, storage, CAPACITY) == 0, "Queue init");
     
     for (int i = 0; i < CAPACITY; i++) {
-        Av1FrameEntry e = { .frame_id = i, .dpb_slot = i, .show_frame = 1, .show_existing_frame = 0 };
+        Av1FrameEntry e = { .frame_id = static_cast<uint32_t>(i), .dpb_slot = i, .show_frame = 1, .show_existing_frame = 0 };
         ASSERT(av1_frame_queue_push(&q, &e) == 0, "Push entry");
     }
     
@@ -43,7 +43,7 @@ static void test_fifo_order(void) {
     av1_frame_queue_init(&q, storage, CAPACITY);
     
     for (int i = 0; i < CAPACITY; i++) {
-        Av1FrameEntry e = { .frame_id = i * 10, .dpb_slot = i, .show_frame = i % 2, .show_existing_frame = 0 };
+        Av1FrameEntry e = { .frame_id = static_cast<uint32_t>(i * 10), .dpb_slot = i, .show_frame = i % 2, .show_existing_frame = 0 };
         av1_frame_queue_push(&q, &e);
     }
